support sd-1.5
