+++ onediff_sd_webui_extensions/compiled_model.py
@@ -1,0 +1,2 @@
+"""oneflow_compiled UNetModel"""
+compiled_unet = None
